修改案例文件格式
