+++ README.md
@@ -1,1 +1,22 @@
-# Domain-PaaS-SDK-Examples-TS
+# 这是一个污水业务中台SDK包使用示例
+
+## 试用申请
+如需试用，请联系我们 [DHI 中国 业务中台](https://online-products.dhichina.cn/) 获取使用许可和认证信息。
+
+## 安装SDK包
+npm install domain-paas-sdk-python
+
+## 示例
+wwtp中是一些sdk的使用示例，可供参考
+
+• 获取全流程水质数据
+参考 wwtp/general_data_example.ts
+
+• 水质模拟预测分析
+参考 wwtp/wq_analysis_example.ts
+
+• 智能反硝化碳源加药
+参考 wwtp/intelligent_denitrification_example.ts
+
+• 工艺模拟实验室
+参考 wwtp/ 参考 wwtp/process_simulation_lab_example.ts
