+++ src/type.ts
@@ -1,0 +1,8 @@
+export interface IToken {
+    access_token: string
+    expires_in: number
+    token_type: string
+    refresh_token: string
+    scope: string
+    error?: string
+}
